Separate engine and httpserver
